// src/lib/types/product.ts
import { ProductCollection, ProductSize, ProductStatus } from "../enums/product.enum";

export interface Product {
  _id: string;
  productStatus: ProductStatus;
  productCollection: ProductCollection;
  productNme: string;
  productPrice: number;
  productLeftCount: number;
  productSize: ProductSize;
  productVolume: number;
  productDesc?: string;
  productImages: string[];
  productViews: number;
  createdAt: Date;
  updateAt: Date;
}
export interface ProductInquiry{
  order:string;
  page: number;
  limit: number;
  productCollection?: ProductCollection;
  search?: string;
}